create theme toggle button automatically on load and toggle theme on user selection
Adding this JS file to an HTML page will create a way to toggle between
light and dark themes. Now, in its minimal form, only two radio buttons
will appear, the first of which is  for the "light" theme and the other
one is for "dark". Selecting the button will apply the corresponding
theme to the current page. The change of theme is consistent across pages
- which means that after selesting "dark" and then going to another page
will not reset the theme to its defualt "light".